feat: rename shared/util to shared/utils
