+++ shared/utils/file-validation.ts
