+++ shared/utils/has-prop.ts
