+++ shared/utils/stringify-safe.ts
