+++ shared/utils/verification.ts
