+++ shared/utils/verified-content.ts
