+++ src/app/modules/myinfo/myinfo.util.ts
@@ -5,7 +5,7 @@ import mongoose from 'mongoose'
 import { err, ok, Result } from 'neverthrow'
 import { v4 as uuidv4, validate as validateUUID } from 'uuid'
 
-import { hasProp } from '../../../../shared/util/has-prop'
+import { hasProp } from '../../../../shared/utils/has-prop'
 import { types as myInfoTypes } from '../../../shared/resources/myinfo'
 import {
   AuthType,
