+++ src/app/modules/verified-content/verified-content.types.ts
@@ -1,6 +1,6 @@
 import { Result } from 'neverthrow'
 
-import { VerifiedKeys } from '../../../../shared/util/verified-content'
+import { VerifiedKeys } from '../../../../shared/utils/verified-content'
 import { AuthType } from '../../../types'
 
 import { MalformedVerifiedContentError } from './verified-content.errors'
