+++ src/app/modules/webhook/webhook.utils.ts
@@ -3,7 +3,7 @@ import { inRange } from 'lodash'
 import moment from 'moment-timezone'
 import { err, ok, Result } from 'neverthrow'
 
-import { stringifySafe } from '../../../../shared/util/stringify-safe'
+import { stringifySafe } from '../../../../shared/utils/stringify-safe'
 import { IWebhookResponse } from '../../../types'
 import { TIMEZONE } from '../../constants/timezone'
 import { randomUniformInt } from '../../utils/random-uniform'
