+++ src/app/modules/webhook/webhook.validation.ts
@@ -1,7 +1,7 @@
 import { promises as dns } from 'dns'
 import ip from 'ip'
 
-import { isValidHttpsUrl } from '../../../../shared/util/url-validation'
+import { isValidHttpsUrl } from '../../../../shared/utils/url-validation'
 import config from '../../config/config'
 
 import { WebhookValidationError } from './webhook.errors'
