+++ src/public/modules/forms/base/components/field-attachment.client.component.js
@@ -4,7 +4,7 @@ const {
   getFileExtension,
   isInvalidFileExtension,
   getInvalidFileExtensionsInZip,
-} = require('../../../../../../shared/util/file-validation')
+} = require('../../../../../../shared/utils/file-validation')
 const { KB, MB } = require('../../../../../shared/constants')
 
 angular.module('forms').component('attachmentFieldComponent', {
