+++ src/public/modules/forms/base/directives/validate-nric.client.directive.js
@@ -1,6 +1,8 @@
 'use strict'
 
-const { isNricValid } = require('../../../../../../shared/util/nric-validation')
+const {
+  isNricValid,
+} = require('../../../../../../shared/utils/nric-validation')
 
 angular.module('forms').directive('validateNric', validateNric)
