+++ src/public/modules/forms/helpers/process-decrypted-content.ts
@@ -7,7 +7,7 @@ import has from 'lodash/has'
 import {
   CURRENT_VERIFIED_FIELDS,
   VerifiedKeys,
-} from '../../../../../shared/util/verified-content'
+} from '../../../../../shared/utils/verified-content'
 import { BasicField, SgidFieldTitle, SPCPFieldTitle } from '../../../../types'
 
 /**
