+++ src/public/services/FormSgSdkService.ts
@@ -1,6 +1,6 @@
 import formSgSdk from '@opengovsg/formsg-sdk'
 
-import { TRANSACTION_EXPIRE_AFTER_SECONDS } from '../../../shared/util/verification'
+import { TRANSACTION_EXPIRE_AFTER_SECONDS } from '../../../shared/utils/verification'
 import { injectedVariables } from '../utils/injectedVariables'
 
 export const FormSgSdk: ReturnType<typeof formSgSdk> = formSgSdk({
